+++ src/Join2-Object_function.ps1
@@ -473,4 +473,29 @@ function Join2-Object
         }
     }    
 }
-Set-Alias -Name join2o -Value Join2-Object -PassThru | Select-Object -Property "DisplayName"
+# set alias
+[String] $tmpAliasName = "join2o"
+[String] $tmpCmdName   = "Join2-Object"
+# is alias already exists?
+if ((Get-Command -Name $tmpAliasName -ErrorAction SilentlyContinue).Count -gt 0){
+    try {
+        if ( (Get-Command -Name $tmpAliasName).CommandType -eq "Alias" ){
+            if ( (Get-Command -Name $tmpAliasName).ReferencedCommand.Name -eq $tmpCmdName ){
+                # pass
+            } else {
+                throw
+            }
+        } else {
+            throw
+        }
+    } catch {
+        Write-Error "Alias ""$tmpAliasName ($((Get-Command -Name $tmpAliasName).ReferencedCommand.Name))"" is already exists. Change alias needed." -ErrorAction Stop
+    } finally {
+        Remove-Variable -Name "tmpAliasName" -Force
+        Remove-Variable -Name "tmpCmdName" -Force
+    }
+} else {
+    Set-Alias -Name $tmpAliasName -Value $tmpCmdName -PassThru | Select-Object -Property "DisplayName"
+    Remove-Variable -Name "tmpAliasName" -Force
+    Remove-Variable -Name "tmpCmdName" -Force
+}